+++ lib/views/screens/home_screen.dart
@@ -1,4 +1,6 @@
 import 'package:app_front/styles/styles.dart';
+import 'package:app_front/views/widgets/course-card.dart';
+import 'package:app_front/views/widgets/custom_app_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_svg/flutter_svg.dart';
@@ -172,155 +174,6 @@ class HomeScreen extends StatelessWidget {
               ),
             ],
           ),
-        ),
-      ),
-    );
-  }
-}
-
-class CustomAppBar extends StatelessWidget implements PreferredSizeWidget {
-  @override
-  Widget build(BuildContext context) {
-    return AppBar(
-      backgroundColor: Colors.white,
-      elevation: 0,
-      title: Row(
-        children: [
-          Padding(
-            padding: const EdgeInsets.only(left: 10),
-            child: Image.asset('assets/images/icons/avatar.jpg', width: 44),
-          ),
-          SizedBox(width: 8),
-          Text(
-            'Username',
-            style: AppFonts.h3,
-            textAlign: TextAlign.center,
-          )
-        ],
-      ),
-      actions: [
-        Padding(
-          padding: EdgeInsets.only(right: 10),
-          child: Image.asset('assets/images/icons/notifications_icon.jpg',
-              width: 44),
-        )
-      ],
-    );
-  }
-
-  @override
-  Size get preferredSize => Size.fromHeight(kToolbarHeight);
-}
-
-class CourseCard extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: EdgeInsets.all(16),
-      margin: EdgeInsets.only(top: 16),
-      width: double.infinity, // Ширина как на макете
-      decoration: BoxDecoration(
-        color: Color(0xFFF0F5FA),
-        borderRadius: BorderRadius.circular(16),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.grey.withOpacity(0.2),
-            spreadRadius: 5,
-            blurRadius: 7,
-            offset: Offset(0, 3),
-          ),
-        ],
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            children: [
-              Image.asset(
-                'assets/images/icons/course_avatar.png',
-                width: 47.53,
-                height: 59.23,
-              ),
-              SizedBox(width: 12),
-              Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    'Graphic Design',
-                    style: AppFonts.titleCourses,
-                  ),
-                  Text(
-                    'Andrew Jhonson',
-                    style: TextStyle(
-                      fontSize: 14,
-                      color: Colors.grey,
-                    ),
-                  ),
-                ],
-              ),
-              Spacer(),
-              Icon(
-                Icons.favorite_border,
-                color: Colors.pink,
-              ),
-            ],
-          ),
-          SizedBox(height: 16),
-          Container(
-            height: 100,
-            width: double.infinity,
-            decoration: BoxDecoration(
-              color: Colors.white,
-              borderRadius: BorderRadius.circular(16),
-            ),
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-              child: TextField(
-                maxLines: null,
-                decoration: InputDecoration(
-                  hintText: 'Enter course details...',
-                  border: InputBorder.none,
-                ),
-              ),
-            ),
-          ),
-          SizedBox(height: 16),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              CourseInfo('3 Chapters'),
-              CourseInfo('30 Lessons'),
-              CourseInfo('3h'),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class CourseInfo extends StatelessWidget {
-  final String text;
-
-  CourseInfo(this.text);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: EdgeInsets.symmetric(vertical: 8, horizontal: 15),
-      width: 110,
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(16),
-        border: Border.all(
-          color: Color(0xFFEBECF2),
-          width: 1.0,
-        ),
-      ),
-      child: Center(
-        child: Text(
-          text,
-          style: TextStyle(fontSize: 14),
         ),
       ),
     );
